+++ src/hooks/usePrevious/data.ts
@@ -1,0 +1,9 @@
+export const parameters = {
+	headers: ['Name', 'Type', 'Description'],
+	data: [['value', '<T>', 'The value to get the previous value from.']],
+}
+
+export const returns = {
+	headers: ['Name', 'Type', 'Description'],
+	data: [['default', '<T>', 'The previous value.']],
+}