+++ src/hooks/useSessionStorage/data.ts
@@ -1,0 +1,19 @@
+export const parameters = {
+	headers: ['Name', 'Type', 'Description'],
+	data: [
+		['key', 'string', 'The key to store the value in the LocalStorage.'],
+		['initalValue', 'string', 'The initial value to store in local storage.'],
+	],
+}
+
+export const returns = {
+	headers: ['Name', 'Type', 'Description'],
+	data: [
+		['ssValue', 'string', 'The value stored in LocalStorage.'],
+		[
+			'setValue',
+			'`(value: string) => void`',
+			'A function to set the value in LocalStorage.',
+		],
+	],
+}