+++ src/hooks/useSetTimeout/index.ts
@@ -1,5 +1,11 @@
 import { useEffect, useState, useRef } from 'react'
 
+/**
+ * Custom hook that sets a timeout and executes a callback function after the specified time has elapsed.
+ * @param initialTime - The initial time in milliseconds before the callback is executed.
+ * @param callback - The callback function to be executed after the specified time has elapsed.
+ * @returns An object containing the `isDone` flag and a `clear` function to cancel the timeout.
+ */
 const useSetTimeout = (
 	initialTime: number,
 	callback: (() => unknown) | (() => Promise<unknown>),
@@ -21,7 +27,17 @@ const useSetTimeout = (
 		}
 	}, [callback, initialTime])
 
-	return isDone
+	const clear = () => {
+		if (timerId.current) {
+			clearTimeout(timerId.current)
+		}
+		setIsDone(true)
+	}
+
+	return {
+		isDone,
+		clear,
+	}
 }
 
 export default useSetTimeout