+++ src/hooks/useToggle/data.ts
@@ -1,0 +1,8 @@
+export const returns = {
+	headers: ['Name', 'Type', 'Description'],
+	data: [
+		['value', 'boolean', 'The value to toggle.'],
+
+		['handleToggle', '() => void', 'The function to toggle the value.'],
+	],
+}