+++ CHANGELOG.md
@@ -1,10 +1,25 @@
 # CHANGELOG
 
+## 1.2.0
+
+### Features
+
+
+- 🎉 Added `clear` function to `useSetInterval` & `useSetTimeout`.
+
+- 🎉 Added new documentation to all hooks.
+
+## 1.1.0
+
+### Features
+
+
+- 🎉 Added Hooks `useToggle` & `useCreen`.
+
 ## 1.0.0
 
 ### Features
 
-#### New features
 
 - 🎉 Added Hooks `useClickOutside` , `useCopyToClipboard` , `useCounter` ,
 `useDebounce`, `useHover` , `useSessionStorage`, `useLocalStorage`, `usePrevious` ,